refactor out add quest form and allow change priority after insert
