+++ imports/ui/Question.jsx
@@ -4,38 +4,30 @@ import { Meteor } from 'meteor/meteor';
 import { Questions } from '../api/questions.js'
  
 export default class Question extends Component {
-
   constructor(props) {
     super(props);
     this.deleteThisQuestion = this.deleteThisQuestion.bind(this);
+    this.updatePriority = this.updatePriority.bind(this);
   }
-
+  updatePriority(e) {
+    Meteor.call('questions.priority.update', this.props.question._id, Number(e.currentTarget.value));
+  }
   deleteThisQuestion() {
     Meteor.call('questions.remove', this.props.question._id);
   }
-
   render() {
-    let priority = '';
-    switch (this.props.question.priority) {
-      case 0:
-        priority = "Urgent";
-        break;
-      case 1:
-        priority = "High";
-        break;
-      case 2:
-        priority = "Medium";
-        break;
-      case 3:
-        priority = "Low";
-        break;
-    }
-    console.log(this.props.question.priority, priority);
+    let priority = this.props.question.priority || '';
     return (
       <li>
         <div className="row">
           <div className="col-lg-2">
-            <span className='priority'>{priority}</span>
+            <select value={priority.toString()}
+                    onChange={this.updatePriority} >
+              <option value='0'>Urgent</option>
+              <option value='1'>High</option>
+              <option value='2'>Medium</option>
+              <option value='3'>Low</option>
+            </select>
           </div>
           <div className="col-lg-4">
             {this.props.question.text}
